Move intermediate outputs
